# tests/snapshots/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "V2ReportActionSuggestion": {
        "properties": {
          "action_id": {
            "default": "",
            "title": "Action Id",
            "type": "string"
          },
          "action_type": {
            "default": "",
            "title": "Action Type",
            "type": "string"
          },
          "can_auto_execute": {
            "default": false,
            "title": "Can Auto Execute",
            "type": "boolean"
          },
          "command": {
            "default": "",
            "title": "Command",
            "type": "string"
          },
          "estimated_cost": {
            "default": "",
            "title": "Estimated Cost",
            "type": "string"
          },
          "priority": {
            "default": 1,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "requires_confirmation": {
            "default": true,
            "title": "Requires Confirmation",
            "type": "boolean"
          },
          "runbook_ref": {
            "default": "",
            "title": "Runbook Ref",
            "type": "string"
          },
          "score": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "title": "V2ReportActionSuggestion",
        "type": "object"
      },
      "V2ReportCompactSummary": {
        "properties": {
          "action_required": {
            "default": false,
            "title": "Action Required",
            "type": "boolean"
          },
          "alert_priority": {
            "default": "",
            "title": "Alert Priority",
            "type": "string"
          },
          "decision": {
            "default": "",
            "title": "Decision",
            "type": "string"
          },
          "primary_action_id": {
            "default": "",
            "title": "Primary Action Id",
            "type": "string"
          },
          "risk_level": {
            "default": "",
            "title": "Risk Level",
            "type": "string"
          },
          "status_line": {
            "default": "",
            "title": "Status Line",
            "type": "string"
          },
          "top_next_command": {
            "default": "",
            "title": "Top Next Command",
            "type": "string"
          },
          "top_retry_hint": {
            "default": "",
            "title": "Top Retry Hint",
            "type": "string"
          }
        },
        "title": "V2ReportCompactSummary",
        "type": "object"
      },
      "V2ReportFailureTaxonomy": {
        "properties": {
          "code": {
            "default": "",
            "title": "Code",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "severity": {
            "default": "",
            "title": "Severity",
            "type": "string"
          },
          "stage": {
            "default": "",
            "title": "Stage",
            "type": "string"
          }
        },
        "title": "V2ReportFailureTaxonomy",
        "type": "object"
      },
      "V2ReportSummaryResponse": {
        "properties": {
          "captured_at": {
            "default": "",
            "title": "Captured At",
            "type": "string"
          },
          "compact": {
            "$ref": "#/components/schemas/V2ReportCompactSummary"
          },
          "failure_taxonomy": {
            "$ref": "#/components/schemas/V2ReportFailureTaxonomy"
          },
          "final_status": {
            "default": "",
            "title": "Final Status",
            "type": "string"
          },
          "latest_pointer": {
            "default": "",
            "title": "Latest Pointer",
            "type": "string"
          },
          "mode": {
            "default": "",
            "title": "Mode",
            "type": "string"
          },
          "next_commands": {
            "items": {
              "type": "string"
            },
            "title": "Next Commands",
            "type": "array"
          },
          "primary_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/V2ReportActionSuggestion"
              },
              {
                "type": "null"
              }
            ]
          },
          "prioritized_actions": {
            "items": {
              "$ref": "#/components/schemas/V2ReportActionSuggestion"
            },
            "title": "Prioritized Actions",
            "type": "array"
          },
          "report_id": {
            "default": "",
            "title": "Report Id",
            "type": "string"
          },
          "report_path": {
            "title": "Report Path",
            "type": "string"
          },
          "retry_hints": {
            "items": {
              "type": "string"
            },
            "title": "Retry Hints",
            "type": "array"
          },
          "source_input": {
            "title": "Source Input",
            "type": "string"
          },
          "subproject_key": {
            "default": "",
            "title": "Subproject Key",
            "type": "string"
          }
        },
        "required": [
          "source_input",
          "report_path"
        ],
        "title": "V2ReportSummaryResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WhereCode Control Center",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reports/v2/summary": {
      "get": {
        "operationId": "get_v2_report_summary_reports_v2_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "subproject",
            "required": false,
            "schema": {
              "default": "stock-sentiment",
              "title": "Subproject",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "report_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Report Id"
            }
          },
          {
            "in": "query",
            "name": "report_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Report Path"
            }
          },
          {
            "in": "query",
            "name": "latest_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Latest Path"
            }
          },
          {
            "in": "query",
            "name": "compact",
            "required": false,
            "schema": {
              "default": false,
              "title": "Compact",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "max_actions",
            "required": false,
            "schema": {
              "default": 3,
              "title": "Max Actions",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Min Score",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "action_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ReportSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get V2 Report Summary"
      }
    },
...
  }
}